requirements.txt
# Core
python-dotenv==1.0.1
PyYAML==6.0.3

# Web automation
selenium==4.24.0

# OCR
pytesseract==0.3.13
Pillow==10.4.0

# Utilidades
requests==2.32.3
tabulate==0.9.0